+++ src/graph.rs
@@ -112,7 +112,15 @@ where
     /// assert_eq!(graph.remove(&Edge::new("a", "c")), false);
     /// ```
     pub fn remove(&mut self, edge: &Edge<T>) -> bool {
-        self.edges.remove(edge)
+        let is_removed = self.edges.remove(edge);
+
+        // Delete the cached objects if the edge was successfully removed because we can't reliably
+        // update them from the new connection alone.
+        if is_removed && self.index.is_some() {
+            self.clear_cache()
+        }
+
+        is_removed
     }
 
     /// Checks if the graph contains an edge.
@@ -870,6 +878,32 @@ mod tests {
 
         // Update the graph with a subset update.
         graph.update_subset("a", &["a", "b"]);
+
+        // Check the cache has been cleared.
+        assert!(graph.index.is_none());
+        assert!(graph.adjacency_matrix.is_none());
+        assert!(graph.degree_matrix.is_none());
+        assert!(graph.laplacian_matrix.is_none());
+    }
+
+    #[test]
+    fn clear_cache_on_remove() {
+        let edge = Edge::new("a", "b");
+        let mut graph = Graph::new();
+        graph.insert(edge.clone());
+
+        // The laplacian requires the computation of the index, the degree matrix and the adjacency
+        // matrix.
+        graph.laplacian_matrix();
+
+        // Check the objects have been cached.
+        assert!(graph.index.is_some());
+        assert!(graph.adjacency_matrix.is_some());
+        assert!(graph.degree_matrix.is_some());
+        assert!(graph.laplacian_matrix.is_some());
+
+        // Update the graph with remove.
+        graph.remove(&edge);
 
         // Check the cache has been cleared.
         assert!(graph.index.is_none());