+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Introduce `Graph::insert_subset` and `Graph::update_subset`
+- Clear internal cache on `Graph::remove`.
+- Introduce `Graph::insert_subset` and `Graph::update_subset`.
 - Manually implement `Default` for `Graph`.
 
 [unreleased]: https://github.com/niklaslong/spectre/compare/v0.2.0...HEAD